pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi>=0.115.11",
    "jinja2>=3.1.6",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.6",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.39",
    "uvicorn>=0.34.0",
]
